+++ src/pages/planos/plano-saude.tsx
@@ -1,10 +1,21 @@
+import { useState, useEffect } from 'react'
 import type { PageProps } from 'gatsby'
 import { graphql } from 'gatsby'
 import type { PlanoSaudeQuery } from '@generated/graphql'
 import { mark } from 'src/sdk/tests/mark'
-import SimpleText from 'src/components/sections/SimpleText/SimpleText'
 import Breadcrumb from 'src/components/sections/Breadcrumb'
-import Section from 'src/components/sections/Section'
+import BuyBox from 'src/components/sections/BuyBox/BuyBox'
+import ExplanationPlan from 'src/components/sections/ExplanationPlan'
+import type { SearchState } from '@faststore/sdk'
+import { SearchProvider, parseSearchState } from '@faststore/sdk'
+import { applySearchState } from 'src/sdk/search/state'
+import { ITEMS_PER_PAGE } from 'src/constants'
+import ProductGallery from 'src/components/sections/ProductGallery'
+import AccordionUp from 'src/components/icons/AccordionUp'
+import AccordionDown from 'src/components/icons/AccordionDown'
+import ScrollToTopButton from 'src/components/sections/ScrollToTopButton'
+
+export type Props = PageProps<PlanoSaudeQuery>
 
 type ItemListType = {
   item: string
@@ -12,12 +23,24 @@ type ItemListType = {
   position: number
 }
 
-export type Props = PageProps<PlanoSaudeQuery>
+const useSearchParams = ({ href }: Location) => {
+  const [params, setParams] = useState<SearchState | null>(null)
+
+  useEffect(() => {
+    const url = new URL(href)
+
+    setParams(parseSearchState(url))
+  }, [href])
+
+  return params
+}
 
 function Page(props: Props) {
   const {
-    data: { allContentfulPlanosTextoSimples },
+    data: { allContentfulPlanos },
   } = props
+
+  const searchParams = useSearchParams(props.location)
 
   const itemListElement: ItemListType[] = [
     {
@@ -34,14 +57,44 @@ function Page(props: Props) {
 
   const title = 'Conheça os planos GoKursos'
 
+  const svgIcons = {
+    svg1: <AccordionUp />,
+    svg2: <AccordionDown />,
+  }
+
+  if (!searchParams) {
+    return null
+  }
+
+  const { galleryTitle } =
+    allContentfulPlanos.nodes[allContentfulPlanos.nodes.length - 1]
+
   return (
-    <Section>
+    <SearchProvider
+      onChange={applySearchState}
+      itemsPerPage={ITEMS_PER_PAGE}
+      {...searchParams}
+    >
       <Breadcrumb breadcrumbList={itemListElement} name={title} />
-      <SimpleText
-        textReceived={allContentfulPlanosTextoSimples}
-        className="text-banner-bottom"
+      <BuyBox
+        nodes={allContentfulPlanos.nodes.filter(
+          (node) => node.slug === '/plano-saude'
+        )}
       />
-    </Section>
+      <ExplanationPlan
+        nodes={allContentfulPlanos.nodes.filter(
+          (node) => node.slug === '/plano-saude'
+        )}
+      />
+
+      <ProductGallery
+        title={title}
+        forceSvg={svgIcons}
+        galleryTitle={galleryTitle}
+      />
+
+      <ScrollToTopButton />
+    </SearchProvider>
   )
 }
 
@@ -52,6 +105,33 @@ export const querySSG = graphql`
         title
         description
         titleTemplate
+      }
+    }
+    allContentfulPlanos(sort: { order: ASC, fields: createdAt }) {
+      nodes {
+        textoBotao
+        titulo
+        preco
+        slug
+        bannerImageMobile {
+          url
+        }
+        bannerImageDesktop {
+          url
+        }
+        compartilhar {
+          url
+        }
+        galleryTitle
+        texto {
+          texto
+        }
+        slug
+      }
+    }
+    allContentfulSignaturePageSubtitle {
+      nodes {
+        subtitle
       }
     }
     allContentfulPlanosTextoSimples {